+++ packages/headless/src/recommendation.index.ts
@@ -1,4 +1,4 @@
-export {Unsubscribe} from '@reduxjs/toolkit';
+export {Unsubscribe, Middleware} from '@reduxjs/toolkit';
 
 export {
   RecommendationEngine,
@@ -8,9 +8,19 @@ export {
   getSampleRecommendationEngineConfiguration,
 } from './app/recommendation-engine/recommendation-engine';
 
+export {CoreEngine, ExternalEngineOptions} from './app/engine';
+export {
+  EngineConfiguration,
+  AnalyticsConfiguration,
+  AnalyticsRuntimeEnvironment,
+} from './app/engine-configuration';
+export {LoggerOptions} from './app/logger';
+export {LogLevel} from './app/logger';
+
 // Actions
 export * from './features/configuration/configuration-actions-loader';
 export * from './features/advanced-search-queries/advanced-search-queries-actions-loader';
+export * from './features/context/context-actions-loader';
 export * from './features/fields/fields-actions-loader';
 export * from './features/pipeline/pipeline-actions-loader';
 export * from './features/search-hub/search-hub-actions-loader';
@@ -19,6 +29,11 @@ export * from './features/recommendation/recommendation-actions-loader';
 export * from './features/recommendation/recommendation-click-analytics-actions-loader';
 
 // Controllers
+export {
+  Controller,
+  buildController,
+} from './controllers/controller/headless-controller';
+
 export {
   RecommendationListOptions,
   RecommendationListProps,
@@ -35,4 +50,7 @@ export {
   buildContext,
 } from './controllers/context/headless-context';
 
+// Miscellaneous
 export {Result} from './api/search/search/result';
+export {HighlightKeyword} from './utils/highlight';
+export {Raw} from './api/search/search/raw';